mainを整頓 Fixes #13
