+++ src/imagenate/main/share.py
@@ -1,0 +1,15 @@
+from argparse import ArgumentParser
+
+from imagenate.libs.enum import CustomEnum
+
+
+class Command(CustomEnum):
+    CONCAT = "concat"
+    CREATE = "create"
+
+
+def create_base_argperser() -> ArgumentParser:
+    argparser = ArgumentParser()
+    argparser.description = "imgage manager"
+    argparser.add_argument("command", type=str, choices=Command.get_values())
+    return argparser
